Further refactoring for I/O of glTF 1.0 and 2.0
The GltfAsset now is a "low-level" data representation that is used
for I/O. The models are created from such an asset, after all
references have been resolved.

Writing a model consists of first creating an asset (either in
default, embedded or binary representation), and then only writing
the asset.

This decouples the model and the file format and representation,
and reduces the amount of version-specific code that is visible
for the clients.