Added getName method to model interfaces
The NamedModelElement interface is a superinterface for the
model interfaces that correspond to a GlTFChildOfRootProperty
in the original glTF asset.

It contains a getName method that allows obtaining the name
that the element has in the glTF input file.

The CameraModel class already had a getName method, which has
been renamed to getInstanceName, to better reflect its usage.