More robust handling of mixed type assets
A GltfReference generally should not contain a data URI,
because it is supposed to be an EXTERNAL reference.
This means that data URIs have to be converted into
buffers manually in the model.
Reading an asset with embedded data and writing it as
binary glTF will NOT convert the embedded data into
binary blobs. This can (or has to be) done manually
by creating a GltfModel, and writing the model as binary.