Large refactoring of camera handling for model building
The cameras are now no longer instantiated multiple times
when the model is created. Instead, instances are created
when the model is rendered. The handling of cameras in the
viewer classes has been simplified significantly by
introducing the RenderedCamera interface, which can either
be an instance of a glTF camera that is attached to a node,
or an external camera.